Do some simple analysis on recent incidents and report.
Dates are hard-coded (to work this week) right now. Campfire support is commented out (no options yet). The code is messy and repetitive.

But, it works for a basic report. I'd like to add some alert volume stuff in there, too. I'm trying to keep it to under the maximum Campfire paste size (20 lines) before a "more" link is added.