Show the volume of alerts in the rotation report.
I'm very sensitive to people getting interrupted in their work, and would like to know if those interruptions are on their way up or down.  I also want to track how often people are being woken up later at night to deal with alerts.

The report now shows the volume of SMS/Phone alerts overall and compared to the last rotation, and also the volume of late-night (10p to 8a) alerts, also compared to the last rotation.

Also started in on some code cleanups thanks to the Hack Arts crew.