Loop over the incident requests to get all data needed.
Previously the incident call only pulled in 100 incidents. Hopefully (!) that would be enough, but just in case, this will make repeated calls (up to 10, for 1000 incidents total) until it finds an incident that is prior to the report period (current rotation and the previous rotation).

The loop sleeps one second between calls for API politeness, so this slows things down if you have a lot of alerts in a report period.