Correct alert count bugs and capture alerts across months.
Assuming that the alerts in a given report (current and previous period) only span two consecutive months, this will now capture all the needed alerts.

Also works around a bug in Chronic; should send the fix to the maintainer.

At this point the rotation report is pretty useful for reporting on the current rotation. Need to add options for output control and for choosing which rotation period to report on.